enrich_service/requirements.txt
celery
Django
django-celery-beat
django-celery-results
djangorestframework
redis
psycopg2
python-decouple
selenium
bs4
linkedin-api==2.0.3
num2words
drf_yasg